+++ R/misc_duplicated_by_names.R
@@ -22,4 +22,3 @@ misc_duplicated_by_names <- function(object, ...) {
   )
   return(duplicated(x = names(object), ...))
 }
-