+++ R/plt_parallel_coordinates.R
@@ -1,9 +1,43 @@
+#' @title plt_parallel_coordinates
+#'
+#' @description Parallel coordinates plot
+#'
+#' @param data A data.table object with the columns containing the parameters to
+#'   be plotted with the parallel coordinates plot.
+#' @param cols A character vector with column names to subset `data` (default:
+#'   `NULL`).
+#' @param color_variable A character. The name of the column to be used to
+#'   colorize the lines of the plot (default: `NULL`).
+#' @param color_args A list with parameters for the color gradient (see
+#'   details).
+#' @param line_jitter A list with the elements `w` and `h` to define a line
+#'   jitter (default: `list(w = 0.04, h = 0.04)`), which are passed further on
+#'   to define the position of `ggplot2::geom_line()`.
+#' @param text_label_size A numeric value to define the size of the text
+#'   annotations (default: `3.5`).
+#'
+#' @details The color gradient of the plotted lines can be defined with a list
+#'   provided to the argument `color_args`. Its default values are
+#'   `alpha = 0.6`, `begin = .1`, `end = .9`, `option = "inferno"`, and
+#'   `direction = 1` and are passed furhter on to
+#'   `ggplot2::scale_color_viridis_c()`.
+#'   The implementation to display categorical variables is still experimental.
+#'
+#' @return Returns a parallel coordinates plot as `ggplot2` object.
+#'
+#' @seealso [ggplot2::scale_color_viridis_c()]
+
+#' @examples
+#' data("iris")
+#' plt_parallel_coordinates(data.table::as.data.table(iris[, -5]))
 #' @export
+#
 plt_parallel_coordinates <- function(
     data,
-    cols,
+    cols = NULL,
     color_variable = NULL,
-    color_args = list(alpha = 0.6, begin = .1, end = .9, option = "inferno", direction = 1),
+    color_args = list(alpha = 0.6, begin = .1, end = .9,
+                      option = "inferno", direction = 1),
     line_jitter = list(w = 0.04, h = 0.04),
     text_label_size = 3.5
 ) {
@@ -16,29 +50,80 @@ plt_parallel_coordinates <- function(
       call. = FALSE
     )
   }
-  stopifnot(is.list(color_args))
-  vec <- colnames(data)[cols]
+
+  # default color parameters
+  color_arg_list <- list(
+    alpha = 0.6,
+    begin = .1,
+    end = .9,
+    option = "inferno",
+    direction = 1
+  )
+  # default line jitter
+  line_jitter_list <- list(
+    w = 0.04,
+    h = 0.04
+  )
+  stopifnot(
+    is.list(line_jitter),
+    intersect(names(line_jitter),
+              names(line_jitter_list)) == names(line_jitter),
+    is.list(color_args),
+    intersect(names(color_args), names(color_arg_list)) == names(color_args),
+    ifelse(
+      is.null(cols), TRUE, intersect(cols, colnames(data)) == cols
+    ),
+    ifelse(
+      is.null(color_variable), TRUE,
+      intersect(color_variable, colnames(data)) == color_variable &&
+        length(color_variable) == 1L
+    ),
+    data.table::is.data.table(data),
+    is.numeric(text_label_size) && length(text_label_size) == 1L
+  )
+
+  # updated default_list
+  color_arg_list <- list.update(
+    main_list = color_arg_list,
+    new_list = color_args
+  )
+
+  # prepare dataframe
+  if (is.null(cols)) {
+    cols <- colnames(data)
+  }
   data_copy <- data.table::copy(data)
   if (!is.null(color_variable)) {
     data_copy <- data_copy[order(get(color_variable), decreasing = TRUE)]
   }
+
+  # round numeric variables
+  num_vars <- colnames(data_copy)[
+    data_copy[, vapply(.SD, is.numeric, logical(1L))]
+  ]
   data_copy[
     ,
-    (vec) := lapply(
+    (num_vars) := lapply(
       X = .SD,
       FUN = function(x) {
         sts_normalize(x) %>%
           round(digits = 3)
       }),
-    .SDcols = vec
+    .SDcols = num_vars
   ]
+
+  # character / factor to integer
+  data_copy <- cat2integer(data_copy)
+
   data_copy$id <- rownames(data)
   data_copy <- data.table::melt.data.table(
     data = data_copy,
     id.vars = c("id", color_variable),
-    measure.vars = vec
+    measure.vars = cols
   )
-  #data_copy[, ("x") := as.integer(get("variable"))]
+  #%data_copy[, ("x") := as.integer(get("variable"))]
+
+  colpos <- which(colnames(data) %in% cols)
 
   data_copy %>%
     ggplot2::ggplot(
@@ -50,22 +135,22 @@ plt_parallel_coordinates <- function(
     ggplot2::geom_line(
       ggplot2::aes_string(group = "id"),
       position = ggplot2::position_jitter(
-        w = line_jitter[["w"]],
-        h = line_jitter[["h"]]
+        w = line_jitter_list[["w"]],
+        h = line_jitter_list[["h"]]
       )
     ) +
     ggplot2::scale_color_viridis_c(
-      option = color_args$option,
-      alpha = color_args$alpha,
-      begin = color_args$begin,
-      end = color_args$end,
-      direction = color_args$direction
+      option = color_arg_list$option,
+      alpha = color_arg_list$alpha,
+      begin = color_arg_list$begin,
+      end = color_arg_list$end,
+      direction = color_arg_list$direction
     ) +
     ggplot2::xlab("Hyperparameter") +
     ggplot2::ylab("Parameter value") +
     geom_label_params(
       data = data,
-      col_pos_vec = cols,
+      col_pos_vec = colpos,
       vjust = -0.5,
       size = text_label_size
     ) +
@@ -74,7 +159,7 @@ plt_parallel_coordinates <- function(
     )
 }
 
-LabelParams <- ggplot2::ggproto(
+LabelParams <- ggplot2::ggproto( # nolint
   `_class` = "LabelParams",
   `_inherit` = ggplot2::GeomText
 )
@@ -88,13 +173,17 @@ geom_label_params <- function(
     parse = FALSE,
     nudge_x = 0,
     nudge_y = 0,
-    label.padding = unit(0.25, "lines"),
-    label.r = unit(0.15, "lines"),
-    label.size = 0.25,
-    na.rm = FALSE,
-    show.legend = NA,
-    inherit.aes = FALSE
+    label.padding = unit(0.25, "lines"), # nolint
+    label.r = unit(0.15, "lines"), # nolint
+    label.size = 0.25, # nolint
+    na.rm = FALSE, # nolint
+    show.legend = NA, # nolint
+    inherit.aes = FALSE # nolint
 ) {
+
+  # character / factor to integer
+  data <- cat2integer(data)
+
   lapply(
     X = seq_len(length(col_pos_vec)),
     FUN = function(x) {
@@ -153,4 +242,28 @@ geom_label_params <- function(
         params = list(na.rm = na.rm, ...),
       )
     })
+}
+
+
+cat2integer <- function(data) {
+  stopifnot(data.table::is.data.table(data))
+  data_copy <- data.table::copy(data)
+  # character / factor to integer
+  cat_vars <- colnames(data_copy)[
+    data_copy[, vapply(.SD, function(x) {
+      is.character(x) || is.factor(x)
+    }, logical(1L))]
+  ]
+  if (length(cat_vars) > 0L) {
+    data_copy[
+      ,
+      (cat_vars) := lapply(
+        X = .SD,
+        FUN = function(x) {
+          as.integer(x) - 1L
+        }),
+      .SDcols = cat_vars
+    ]
+  }
+  return(data_copy)
 }