+++ R/stats_normalize.R
@@ -1,4 +1,0 @@
-#' @export
-sts_normalize <- function(x, na.rm = TRUE) {
-  return((x - min(x)) / (max(x) - min(x)))
-}