+++ tests/testthat/test-mlh.R
@@ -4,8 +4,8 @@ test_that(
 
     fold_list <- list(
       "Fold1" = setdiff(seq_len(100), 1:33),
-      "Fold2" = setdiff(seq_len(100),66:100),
-      "Fold3" = setdiff(seq_len(100),34:65)
+      "Fold2" = setdiff(seq_len(100), 66:100),
+      "Fold3" = setdiff(seq_len(100), 34:65)
     )
     l1 <- mlh_outsample_row_indices(fold_list, 100)
     expect_type(l1, "list")